+++ precapston/src/main/java/com/precapston/precapston/domain/posts/Member.java
@@ -1,34 +1,34 @@
-//package com.precapston.precapston.domain.posts;
-//
-//import jakarta.persistence.*;
-//import lombok.Builder;
-//import lombok.Getter;
-//import lombok.NoArgsConstructor;
-//
-//
-//@Getter
-//@NoArgsConstructor
-//@Entity
-//public class Member {
-//
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//
-//    @Column(length = 20, nullable = false)
-//    private String name;
-//
-//    @Column(length = 20, nullable = false)
-//    private String birth;
-//
-//    @Column(length=15,nullable = false)
-//    private String phone_num;
-//
-//    @Column(length=20, nullable = false)
-//    private String login_id;
-//
-//    @Column(length=20, nullable = false)
-//    private String login_pw;
-//
-//    @Column
-//}
+package com.precapston.precapston.domain.posts;
+
+import jakarta.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+
+@Getter
+@NoArgsConstructor
+@Entity
+public class Member {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 20, nullable = false)
+    private String name;
+
+    @Column(length = 20, nullable = false)
+    private String birth;
+
+    @Column(length=15,nullable = false)
+    private String phone_num;
+
+    @Column(length=20, nullable = false)
+    private String login_id;
+
+    @Column(length=20, nullable = false)
+    private String login_pw;
+
+
+}
